+++ code/FinanceManager.Infrastructure/Repositories/Account/Entry/BondEntryRepository.cs
@@ -155,6 +155,7 @@ public class BondEntryRepository(AppDbContext context) : IBondAccountEntryReposi
 
     public async Task<BondAccountEntry?> Get(int accountId, int entryId) => await context.BondEntries
             .AsNoTracking()
+            .Include(entry => entry.Labels)
             .FirstOrDefaultAsync(x => x.AccountId == accountId && x.EntryId == entryId);
 
     public async Task<int> GetCount(int accountId) => await context.BondEntries.AsNoTracking().CountAsync(x => x.AccountId == accountId);